+++ src/main/java/com/example/streaming/service/ContentPostServiceImpl.java
@@ -1,9 +1,9 @@
-package com.example.streaming.service.impl;
+package com.example.streaming.service;
 
 import com.example.streaming.domain.ContentPost;
-import com.example.streaming.repository.ContentPostRepository;
-import com.example.streaming.service.ContentPostService;
-import com.example.streaming.dto.ContentPostCreate;
+import com.example.streaming.service.port.ContentPostRepository;
+import com.example.streaming.controller.port.ContentPostService;
+import com.example.streaming.controller.request.ContentPostCreate;
 import lombok.Builder;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
